Add OSX support
- Fixes typo in INKSCAPE_DIR name
- makes INKSCAPE_DIR location conditional on uname. This will not
  work on windows; if uname is unavailable that's as good a windows
  check as any.
- switches from python to python3. python3 always has that symlink,
  on Linux too, but in Homebrew the python symlink may not be in your path
  see https://docs.brew.sh/Homebrew-and-Python - it's in libexec

Handling of spaces in paths isn't great, because it's a Makefile.
The space in "Application Support" is escaped but spaces in $HOME
are not. This isn't really fixable in Make, and the windows install
is going to be awkward with both Make and bash needed; might be
that going python-only is the answer.